replicate font-size preference on UserProvider
