+++ apps/meteor/client/providers/UserProvider/UserProvider.tsx
@@ -13,6 +13,7 @@ import { afterLogoutCleanUpCallback } from '../../../lib/callbacks/afterLogoutCl
 import { useIsEnterprise } from '../../hooks/useIsEnterprise';
 import { useReactiveValue } from '../../hooks/useReactiveValue';
 import { createReactiveSubscriptionFactory } from '../../lib/createReactiveSubscriptionFactory';
+import { useCreateFontStyleElement } from '../../views/account/accessibility/hooks/useCreateFontStyleElement';
 import { useEmailVerificationWarning } from './hooks/useEmailVerificationWarning';
 import { useLDAPAndCrowdCollisionWarning } from './hooks/useLDAPAndCrowdCollisionWarning';
 
@@ -69,6 +70,8 @@ const UserProvider = ({ children }: UserProviderProps): ReactElement => {
 
 	const { data: license } = useIsEnterprise();
 	const setUserPreferences = useEndpoint('POST', '/v1/users.setPreferences');
+
+	useCreateFontStyleElement(user?.settings?.preferences?.fontSize);
 
 	const loginMethod: LoginMethods = (isLdapEnabled && 'loginWithLDAP') || (isCrowdEnabled && 'loginWithCrowd') || 'loginWithPassword';
